(PC-10080) Add SectionWithDivider ui component
