+++ src/features/cheatcodes/pages/AppComponents/AppComponents.tsx
@@ -15,10 +15,8 @@ import { UseNavigationType } from 'features/navigation/RootNavigator'
 import { CallToAction } from 'features/offer/components'
 import { BeneficiaryCeilings } from 'features/profile/components/BeneficiaryCeilings'
 import { NonBeneficiaryHeader } from 'features/profile/components/NonBeneficiaryHeader'
-import { accordionStyle } from 'features/profile/components/reusables'
 import { SelectionLabel } from 'features/search/atoms/SelectionLabel'
 import { CATEGORY_CRITERIA } from 'features/search/enums'
-import { analytics } from 'libs/analytics'
 import { mapCategoryToIcon } from 'libs/parsers'
 import { AccordionItem } from 'ui/components/AccordionItem'
 import { Badge } from 'ui/components/Badge'
@@ -46,6 +44,7 @@ import { ModalHeader } from 'ui/components/modals/ModalHeader'
 import { useModal } from 'ui/components/modals/useModal'
 import { RadioButton } from 'ui/components/RadioButton'
 import { SectionRow } from 'ui/components/SectionRow'
+import { SectionWithDivider } from 'ui/components/SectionWithDivider'
 import { useSnackBarContext } from 'ui/components/snackBar/SnackBarContext'
 import { SocialNetworkCard } from 'ui/components/SocialNetworkCard'
 import { SocialNetworkIconsMap, SocialNetwork } from 'ui/components/socials/types'
@@ -163,23 +162,10 @@ export const AppComponents: FunctionComponent = () => {
 
   const { goBack } = useNavigation<UseNavigationType>()
 
-  const someID = 1234
-
   return (
     <StyledScrollView>
       <Spacer.TopScreen />
       <ModalHeader title="App components" leftIcon={ArrowPrevious} onLeftIconPress={goBack} />
-
-      {/* AccordionItem */}
-      <AccordionItem
-        title="Accordion Item"
-        defaultOpen={true}
-        onOpenOnce={() => analytics.logConsultWithdrawal(someID)}
-        onOpen={() => analytics.logConsultWithdrawal(someID)}
-        titleStyle={accordionStyle.title}
-        bodyStyle={accordionStyle.body}>
-        <View />
-      </AccordionItem>
 
       {/* Typos */}
       <AccordionItem title="Typos">
@@ -293,6 +279,11 @@ export const AppComponents: FunctionComponent = () => {
       </AccordionItem>
 
       <Divider />
+
+      {/* Sections */}
+      <SectionWithDivider visible margin>
+        <View />
+      </SectionWithDivider>
 
       {/* Switches */}
       <AccordionItem title="Switches">
