+++ src/features/venue/fixtures/venueResponseSnap.ts
@@ -1,7 +1,7 @@
 import { VenueResponse } from 'api/gen'
 
 export const venueResponseSnap: VenueResponse = {
-  id: 5576,
+  id: 5543,
   name: 'Le Petit Rintintin 1',
   latitude: 48.87004,
   longitude: 2.3785,
